ApplicationLatte: Template divided into hierarchy LatteTemplate > Template (dynamic) + StrictTemplate
